requirements-tests.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements-tests.txt requirements-tests.in
#
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
bandit==1.6.0
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
coverage==4.5.3
flake8==3.5.0
funcparserlib==0.3.6      # via mockldap
gitdb2==2.0.5             # via gitpython
gitpython==2.1.11         # via bandit
idna==2.8                 # via requests
importlib-metadata==0.18  # via pluggy, pytest
mccabe==0.6.1             # via flake8
mock==3.0.5
mockldap==0.3.0
more-itertools==7.0.0     # via pytest
nose==1.3.7
packaging==19.0           # via pytest
pathlib2==2.3.3           # via pytest
pbr==5.3.0                # via stevedore
pluggy==0.12.0            # via pytest
py==1.8.0                 # via pytest
pyasn1-modules==0.2.5     # via python-ldap
pyasn1==0.4.5             # via pyasn1-modules, python-ldap
pycodestyle==2.3.1        # via flake8
pyflakes==1.6.0
pyldap==3.0.0.post1       # via mockldap
pyparsing==2.4.0          # via packaging
pytest-mock==1.10.4
pytest==4.6.3
python-ldap==3.2.0        # via pyldap
pyyaml==5.1.1
requests-mock==1.6.0
requests==2.22.0          # via requests-mock
six==1.12.0               # via bandit, mock, packaging, pathlib2, pytest, requests-mock, stevedore
smmap2==2.0.5             # via gitdb2
stevedore==1.30.1         # via bandit
urllib3==1.25.3           # via requests
wcwidth==0.1.7            # via pytest
zipp==0.5.1               # via importlib-metadata
